requirements.txt
Pillow
screeninfo
py-cpuinfo
psutil